move every encoder-related package behind encoder
