+++ exporter/elasticsearchexporter/internal/encoder/ecs/attributes.go
