+++ exporter/elasticsearchexporter/internal/encoder/ecs/attributes_test.go
