+++ exporter/elasticsearchexporter/internal/encoder/otel/scope.go
